+++ lib/searchkick/query.rb
@@ -871,6 +871,11 @@ module Searchkick
     end
 
     def where_filters(where)
+      if where.respond_to?(:permitted?) && !where.permitted?
+        # TODO check in more places
+        Searchkick.warn("Passing unpermitted parameters will raise an exception in Searchkick 5")
+      end
+
       filters = []
       (where || {}).each do |field, value|
         field = :_id if field.to_s == "id"